src/test/java/com/thoughtworks/api/infrastructure/repositories/ProductRepositoryTest.java
package com.thoughtworks.api.infrastructure.repositories;

import com.thoughtworks.api.infrastructure.core.Product;
import com.thoughtworks.api.support.DatabaseTestRunner;
import com.thoughtworks.api.support.TestHelper;
import org.junit.Test;
import org.junit.runner.RunWith;

import javax.inject.Inject;
import java.util.HashMap;
import java.util.Map;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertThat;

/**
 * Created by syzhang on 7/17/16.
 */
@RunWith(DatabaseTestRunner.class)
public class ProductRepositoryTest {
    @Inject
    ProductRepository productRepository;

    @Test
    public void should_create_and_get_product(){


        Product product = productRepository.createProduct(TestHelper.product("apple"));

        assertThat(product.getName(), is("apple"));
        assertThat(product.getDescription(), is("red apple"));
        assertEquals(product.getPrice(), 1.1, 0.01);

    }
}
